agent/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.3
python-dotenv==1.0.1
python-consul==1.1.0
pydantic==2.6.1
pydantic-settings==2.1.0
aiohttp==3.9.3
cryptography==42.0.5 